+++ src/styles/Newbook.scss
@@ -22,7 +22,6 @@ $padding: 5px 10px;
     justify-content: flex-end;
     gap: 90px;
     padding: 5px;
-    
 
     .add-book {
       width: 45%;
@@ -66,6 +65,7 @@ $padding: 5px 10px;
       cursor: pointer;
     }
   }
+
   .message {
     color: red;
   }
