+++ src/styles/button.scss
@@ -1,0 +1,13 @@
+$primaryColor: rgb(46, 164, 255);
+
+.delete-btn {
+  padding: 10px 20px;
+  color: #fff;
+  background-color: $primaryColor;
+  border: none;
+  border-radius: 5px;
+}
+
+.delete-btn:hover {
+  cursor: pointer;
+}
